requirements.txt
fastapi
uvicorn[standard]
httpx
sqlmodel
dataclasses-json
pydantic
aiofiles
python-multipart
openai
faiss-cpu
pdfminer.six
beautifulsoup4
readability-lxml
tqdm
python-dotenv
pydub
requests
aiohttp
markdown
requests
